fix: render metrics and report tables on Result page
- MixSession.Bpm and MixSessionDto.Bpm: float? -> double? (SQL float is
  double-precision; EF Core GetFloat threw InvalidCastException)
- AnalysisAgent: GetSingle() -> (float)GetDouble() for all numeric fields
  from Python JSON (Python always emits double-precision)
- Result.razor: _before/_after switched from JsonElement? to
  Dictionary<string,double>; ParseAnalysis deserializes via
  Dictionary<string,JsonElement>, GetFloat reads from dict directly

+++ AudioPipeline.Shared/DTOs/MixSessionDto.cs
@@ -6,7 +6,7 @@ public class MixSessionDto
     public string JobId { get; set; } = string.Empty;
     public string InputFile { get; set; } = string.Empty;
     public string? Genre { get; set; }
-    public float? Bpm { get; set; }
+    public double? Bpm { get; set; }
     public string? Key { get; set; }
     public string Status { get; set; } = string.Empty;
     public string? BlocksRun { get; set; }
